Move PredicateTuple
